refactor(mergeScripts): rewrite typescript
