+++ src/index.ts
@@ -63,7 +63,7 @@ const modules = {
     custom: () => interop(import('./_modules/custom')),
     deduplicateAttributeValues: () => interop(import('./_modules/deduplicateAttributeValues')),
     // example: () => import('./_modules/example.mjs'),
-    mergeScripts: () => import('./_modules/mergeScripts.mjs'),
+    mergeScripts: () => interop(import('./_modules/mergeScripts.js')),
     mergeStyles: () => interop(import('./_modules/mergeStyles.js')),
     minifyConditionalComments: () => interop(import('./_modules/minifyConditionalComments')),
     minifyCss: () => import('./_modules/minifyCss.mjs'),
